move `reconcileStatefulSet` to `statefulset.go`
